+++ src/Buttr.Core/Configurables/IConfigurable.cs
@@ -4,5 +4,6 @@ namespace Buttr.Core {
     public interface IConfigurable<TConcrete> {
         IConfigurable<TConcrete> WithConfiguration(Func<TConcrete, TConcrete> configuration);
         IConfigurable<TConcrete> WithFactory(Func<TConcrete> factory);
+        IConfigurable<TConcrete> As<TAlias>();
     }
 }