+++ src/Buttr.Core/Configurables/IConfigurableCollection.cs
@@ -4,5 +4,6 @@ namespace Buttr.Core {
     public interface IConfigurableCollection {
         public IConfigurableCollection WithConfiguration<TConcrete>(Func<TConcrete, TConcrete> configuration);
         public IConfigurableCollection WithFactory<TConcrete>(Func<TConcrete> factory);
+        public IConfigurableCollection As<TConcrete, TAlias>();
     }
 }