+++ src/Buttr.Core/Containers/Interfaces/IDIContainer.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Buttr.Core {
     public interface IDIContainer<in TID> : IDisposable {
@@ -7,9 +8,10 @@ namespace Buttr.Core {
         T Get<T>(TID id);
         bool TryGet<T>(TID id, out T value);
     }
-    
+
     public interface IDIContainer : IDisposable {
         T Get<T>();
         bool TryGet<T>(out T value);
+        IEnumerable<T> All<T>();
     }
 }