+++ src/Buttr.Core/Registry/Registration.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Buttr.Core {
     internal sealed class Registration {
@@ -6,11 +7,17 @@ namespace Buttr.Core {
         internal Type PrimaryKey { get; }
         internal Type ConcreteType { get; }
         internal bool IsHidden { get; }
+        internal HashSet<Type> Aliases { get; private set; }
 
         internal Registration(Type primaryKey, Type concreteType, bool isHidden) {
             PrimaryKey = primaryKey;
             ConcreteType = concreteType;
             IsHidden = isHidden;
+        }
+
+        internal void AddAlias(Type aliasType) {
+            Aliases ??= new HashSet<Type>();
+            Aliases.Add(aliasType);
         }
     }
 }