+++ src/Buttr.Core/Resolvers/Static/HiddenStaticTransientResolver.cs
@@ -23,7 +23,7 @@ namespace Buttr.Core {
 
             m_Transient.Dispose();
             m_Transient = null;
-            ApplicationRegistry.Remove(m_Registration.PrimaryKey);
+            ApplicationRegistry.Remove(m_Registration);
         }
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithConfiguration(Func<TConcrete, TConcrete> configuration) {
@@ -33,6 +33,11 @@ namespace Buttr.Core {
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithFactory(Func<TConcrete> factory) {
             m_Factory = factory;
+            return this;
+        }
+
+        IConfigurable<TConcrete> IConfigurable<TConcrete>.As<TAlias>() {
+            AliasSupport.StageApplicationAlias<TConcrete, TAlias>(m_Registration);
             return this;
         }
     }
@@ -59,7 +64,7 @@ namespace Buttr.Core {
 
             m_Transient.Dispose();
             m_Transient = null;
-            ApplicationRegistry.Remove(m_Registration.PrimaryKey);
+            ApplicationRegistry.Remove(m_Registration);
         }
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithConfiguration(Func<TConcrete, TConcrete> configuration) {
@@ -69,6 +74,11 @@ namespace Buttr.Core {
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithFactory(Func<TConcrete> factory) {
             m_Factory = factory;
+            return this;
+        }
+
+        IConfigurable<TConcrete> IConfigurable<TConcrete>.As<TAlias>() {
+            AliasSupport.StageApplicationAlias<TConcrete, TAlias>(m_Registration);
             return this;
         }
     }