+++ src/Buttr.Core/Resolvers/Static/StaticTransientResolver.cs
@@ -17,6 +17,7 @@ namespace Buttr.Core {
             m_Registration.Resolver = m_Transient;
             ApplicationRegistry.Register(m_Registration);
             Application<TConcrete>.Set(m_Transient);
+            AliasSupport.SetApplicationForAliases(m_Registration, m_Transient);
         }
 
         public void Dispose() {
@@ -25,7 +26,7 @@ namespace Buttr.Core {
             m_Transient.Dispose();
             m_Transient = null;
 
-            ApplicationRegistry.Remove(m_Registration.PrimaryKey);
+            ApplicationRegistry.Remove(m_Registration);
             Application<TConcrete>.Set(null);
         }
 
@@ -36,6 +37,11 @@ namespace Buttr.Core {
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithFactory(Func<TConcrete> factory) {
             m_Factory = factory;
+            return this;
+        }
+
+        IConfigurable<TConcrete> IConfigurable<TConcrete>.As<TAlias>() {
+            AliasSupport.StageApplicationAlias<TConcrete, TAlias>(m_Registration);
             return this;
         }
     }
@@ -56,6 +62,7 @@ namespace Buttr.Core {
             m_Registration.Resolver = m_Transient;
             ApplicationRegistry.Register(m_Registration);
             Application<TAbstract>.Set(m_Transient);
+            AliasSupport.SetApplicationForAliases(m_Registration, m_Transient);
         }
 
         public void Dispose() {
@@ -63,7 +70,8 @@ namespace Buttr.Core {
 
             m_Transient.Dispose();
             m_Transient = null;
-            ApplicationRegistry.Remove(m_Registration.PrimaryKey);
+            AliasSupport.ClearApplicationForAliases(m_Registration);
+            ApplicationRegistry.Remove(m_Registration);
             Application<TAbstract>.Set(null);
         }
 
@@ -74,6 +82,11 @@ namespace Buttr.Core {
 
         IConfigurable<TConcrete> IConfigurable<TConcrete>.WithFactory(Func<TConcrete> factory) {
             m_Factory = factory;
+            return this;
+        }
+
+        IConfigurable<TConcrete> IConfigurable<TConcrete>.As<TAlias>() {
+            AliasSupport.StageApplicationAlias<TConcrete, TAlias>(m_Registration);
             return this;
         }
     }